requirements.txt
pybind11==2.9.1
absl-py==1.0.0
aiodns==3.0.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.0.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
DAWG-Python==0.7.2
dill==0.3.4
docopt==0.6.2
fasttext-wheel==0.9.2
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gql==2.0.0
graphql-core==2.3.2
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
multidict==6.0.2
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
pandas==1.4.0
prometheus-client==0.7.1
promise==2.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycparser==2.21
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.2
python-function-model==0.0.1
pytz==2021.3
requests-oauthlib==1.3.1
rsa==4.8
Rx==1.6.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
python_function_model==0.0.1
prometheus_client==0.7.1
gql==2.0.0
aiohttp~=3.7.1
aiohttp-requests==0.1.3
requests~=2.26.0
PyYAML~=5.4.1
Flask~=2.0.1
pymorphy2~=0.9.1
flask_json